product form refactoring complete; form itself moved to components, formpage renamed to formmaster, fetch calls moved to productmanager, routes and indexes updated accordingly
