+++ src/components/form/index.js
@@ -3,3 +3,4 @@ export { default as RegisterForm } from "./register";
 export { default as PaymentForm } from "./payment";
 export { default as UserForm } from "./user";
 export {default as SearchForm} from "./searchForm"
+export {default as ProductForm} from "./product"
