+++ src/pages/products/form/testCharacters.js
