+++ src/pages/products/index.js
@@ -1,1 +1,2 @@
 export {default as ProductDetails} from "./productDetails"
+export {default as ProductFormMaster} from "./form/Master"
